修改代理余额的单位到分---类型int到double

+++ src/main/java/com/mtf/admin/common/vo/PerformanceVO.java
@@ -7,6 +7,6 @@ import java.io.Serializable;
 @Data
 public class PerformanceVO implements Serializable {
     private Integer month;
-    private Long levelOneMoney;
-    private Long levelTwoMoney;
+    private Integer levelOneMoney;
+    private Integer levelTwoMoney;
 }
